So it turns out tr1's unordered_map is not thread safe.
The only way to avoid undefined behavior in the collision iterator is to
serialize access to it.  And that defeats the whole purpose of multi-threaded
collision detection.  What I need is to be able to let more than one item
access the collisions list, so long as they're not both trying to write to the
same collision at the same time.  That's supposed to be what the
collisions_mutexes map is for, but even if multiple threads aren't accessing
the same collision, they still can't simultaneously access the collisions list,
either.

Very disappointing.